pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-radar"
version = "0.1.0"
description = "Daily AI-assisted personal event radar"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "openai>=2.0.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "RapidFuzz>=3.9.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
event-radar = "event_radar.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
